requirements.txt
torch
pytorch-lightning
accelerate
transformers
wandb
webdataset
peft
einops
pillow
braceexpand